+++ packages/server-hono/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @voltagent/server-hono
 
+## 1.0.15
+
+### Patch Changes
+
+- [`ca6160a`](https://github.com/VoltAgent/voltagent/commit/ca6160a2f5098f296729dcd842a013558d14eeb8) Thanks [@omeraplak](https://github.com/omeraplak)! - fix: updates endpoint
+
+- Updated dependencies [[`ca6160a`](https://github.com/VoltAgent/voltagent/commit/ca6160a2f5098f296729dcd842a013558d14eeb8)]:
+  - @voltagent/server-core@1.0.14
+
+## 1.0.14
+
+### Patch Changes
+
+- [#629](https://github.com/VoltAgent/voltagent/pull/629) [`3e64b9c`](https://github.com/VoltAgent/voltagent/commit/3e64b9ce58d0e91bc272f491be2c1932a005ef48) Thanks [@omeraplak](https://github.com/omeraplak)! - feat: add memory observability
+
+- Updated dependencies [[`3e64b9c`](https://github.com/VoltAgent/voltagent/commit/3e64b9ce58d0e91bc272f491be2c1932a005ef48)]:
+  - @voltagent/server-core@1.0.13
+  - @voltagent/core@1.1.22
+
 ## 1.0.13
 
 ### Patch Changes